requirements.txt
asgiref==3.3.1
certifi==2020.6.20
Django==3.1.7
numpy==1.19.2
pytz==2020.1
scipy==1.5.3
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
pillow==8.0.0
psycopg2-binary
django-storages[azure]
opencensus-ext-azure
